requirements.txt
Django==5.0.1
djangorestframework==3.14.0
neomodel==5.2.1
firebase-admin==6.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
networkx==3.2.1